Refatorando (#4)
* Relocate scripts and source files from .github to src

All scripts and source files have been relocated from .github directory to src directory. This also led to the revision of paths in update_bucket.yml, README.md, and .gitignore respectively. Additionally, the directory for Dependabot has been adjusted to the new location for the pip ecosystem.

* Add new certificates and update README commands

This commit adds new certificate files to the project: 'client.pem' and 'projeto.corp.crt'. It also corrects the script execution commands in the README file. The corrected commands now point to the appropriate script directory and use a revised naming convention for Brazil.