Update GitHub workflow action path
The path for the "Use action defined in action.yml" step in the update_bucket workflow has been modified. Previously, it was directly referencing the main branch of the "bucket_python" repository. The revised path now correctly points to the "action.yml" file inside the ".github/workflows/" directory of the same repository.