Update action path in GitHub workflow
The commit revises the path used for the 'Synchronize Certificates' action in the 'update_bucket.yml' GitHub workflow. Instead of pointing to 'lisboajeff/bucket_python@main', it now points to 'lisboajeff/bucket_python/.github/actions@main'. This change ensures that the correct actions are being used when the workflow is run.