Update GitHub workflow for certificate synchronization
The GitHub workflow for synchronizing certificates to S3 has been updated to now trigger on pull request events, instead of manually dispatch. The workflow includes a new job for checking if the commit message follows the required template. This changes are aimed at making the certificate synchronization process more automated, efficient and standard-compliant.