country Brasil environment dev certificates [client] (#6)
* Update commit message check to title message check

This commit modifies the .github/workflows/updated_bucket.yml file to change the validation from commit messages to title messages. This is to align with our new policy of using pull request titles, instead of commit messages, as primary communication tool.

* Fix regex check in update_bucket workflow

The title regex check in the update_bucket.yml GitHub Actions workflow was incorrectly being compared with a string "TITLE_REGEX" instead of the actual regex stored in the variable. This change corrects this, ensuring the validation check runs properly by comparing the title with the defined regex.

* Fix pull request title fetching in GitHub action

Updated GitHub action script to properly fetch the title of PR during workflow. This fix primarily ensures the title string is reassigned correctly, preventing incorrect matching with its regular expression.

* Update title message display in update_bucket.yml

The title message in GitHub workflow 'update_bucket.yml' has been updated to correctly display the title of the pull request. Previously, it was printing only the string "TITLE", and did not reflect the actual title of the pull request.

* Update regex and error messaging in update_bucket workflow

The previous regex for checking the pull request's title in the update_bucket GitHub action has been modified. This commit simplifies the pattern and accommodates more character types. In addition, the title message is now included in the error message when the title does not match the required template, thereby improving clarity.

* Remove fetch-depth from checkout action

The fetch-depth parameter in the action checkout@v2 within the update_bucket workflow file has been removed. This change simplifies the Checkout code step, as fetch-depth of 0 is no longer necessary for the operation of the

* Update GitHub Action to extract country and environment

Modified the GitHub workflow to extract the country and environment values from the commit message instead of the GitHub event inputs. This update will ensure the correct certificates are synced based on the metadata in the commit message.

* Update Github workflow environment variable references

The commit updates the environment variable references in the "update_bucket" Github workflow. Instead of accessing variables through "env", the commit changes the reference to use a direct "$" prefix. This simplifies and harmonizes variable usage within the workflow configuration.

* Refactor update_bucket workflow to output variables

The update_bucket workflow has been refactored to use the "::set-output" command for output variables, instead of using "$GITHUB_ENV". This change enhances the clarity of the code by providing explicit output values, improving readability, and improving maintainability.

* Update output method in Github workflow

The commit changes the way the 'country' and 'environment' variables are output in the update_bucket.yml GitHub workflow. Instead of using the echo command to display the values, the variables are now appended to the GITHUB_OUTPUT file for further usage.

* Update output variables in GitHub workflow

The commit updates the 'check-squash-commit-message' section in the GitHub workflow configuration file. The source of the 'country' and 'environment' output variables has been updated from 'set_vars' steps to 'check' steps. A new 'check' step has also been added to validate the pull request title format.

* Update job name and fix references in GitHub workflow

Renamed the GitHub Actions job from 'check-squash-commit-message' to 'verifier' and updated all relevant references accordingly. Also added an echo command to print the GitHub output for better debugging.

* Update workflow dependency in update_bucket.yml

The dependency in the 'sync_certificates' section of the GitHub workflow has been updated. The action now requires 'verifier' to be completed before it can run, instead of 'check-squash-commit-message'. This change has been implemented to ensure the correct workflow sequence.

* Replace echo with cat in GitHub workflow

The echo command was replaced with the cat command in the update_bucket workflow. This is to address an issue where the contents of the GITHUB_OUTPUT environment variable were not being properly displayed. The cat command provides a more robust solution for outputting its contents.

* Update GitHub workflow variables

The commit updates the variable references in the 'update_bucket.yml' GitHub workflow. The update corrects the previous usage of 'needs.verifier.outputs' to the more accurate 'verifier.outputs', ensuring correct fetching and usage of 'country', 'environment', and 'certificates' values.

* Update references in GitHub action

The references have been changed in the `update_bucket.yml` GitHub workflow file. The `verifier.outputs` references were updated to `needs.verifier.outputs`. This adjustment ensures that the correct output from the "verifier" action is used in subsequent steps within the workflow.

* Update output format in update_bucket workflow

The echo commands in the GitHub action workflow file, update_bucket.yml, were altered. Instead of using curly braces, the outputs for the "country" and "environment" variables are now written in a more standard format. This change improves the readability and compatibility of the log output.