country Brasil environment dev certificates [client] (#8)
* Remove project description from README

The README.md file was refactored by removing all the detailed project description, prerequisites, features, and how-to-use instructions. This simplifies the README to just display a title, making it more concise and easily maintainable. Future documentation will be placed in relevant directories.

* Update pull request title validation regex

The pull request title validation regular expression in the update_bucket GitHub action workflow has been updated. The new regex now correctly accounts for potential hyphenated values in the certificates field.

* Update regex pattern in update_bucket.yml

The update modifies the title validation regex pattern used in the update_bucket.yml GitHub workflow. Specifically, an escape character is added before the hyphen in the regex pattern, which ensures proper interpretation of the character sequence

* Update regex pattern in update_bucket.yml

The update modifies the title validation regex pattern used in the update_bucket.yml GitHub workflow. Specifically, an escape character is

* Update GitHub workflow and regex pattern

The GitHub workflow file 'update_bucket.yml' has been updated to use different arguments in the 'sync_certificates' function. This includes changes to use a 'base_path', 'description', 'configuration', and 'expression_file_paths' instead of 'country', 'environment', and 'certificates'. Additionally, the regex pattern used in the workflow to check the pull request title has been corrected to improve the matching functionality.

* Remove 'TRUSTSTORE' and 'TLS' from config

The configurations 'TRUSTSTORE' and 'TLS' were removed from the Brasil/dev/config.env file. These changes are considered necessary for the new application's environment setup and reducing unnecessary complexity.

* Update reference in GitHub workflow

The reference to an action file in the GitHub workflow named 'update_bucket.yml' was updated. Specifically, the line was changed from using the 'refatorando' branch to now use the 'main' branch in the repository. This change ensures that the workflow uses the latest stable version of the action.