country Brasil environment dev update (#9)
* Remove project description from README

The README.md file was refactored by removing all the detailed project description, prerequisites, features, and how-to-use instructions. This simplifies the README to just display a title, making it more concise and easily maintainable. Future documentation will be placed in relevant directories.

* Update pull request title validation regex

The pull request title validation regular expression in the update_bucket GitHub action workflow has been updated. The new regex now correctly accounts for potential hyphenated values in the certificates field.

* Update regex pattern in update_bucket.yml

The update modifies the title validation regex pattern used in the update_bucket.yml GitHub workflow. Specifically, an escape character is added before the hyphen in the regex pattern, which ensures proper interpretation of the character sequence

* Update regex pattern in update_bucket.yml

The update modifies the title validation regex pattern used in the update_bucket.yml GitHub workflow. Specifically, an escape character is

* Update GitHub workflow and regex pattern

The GitHub workflow file 'update_bucket.yml' has been updated to use different arguments in the 'sync_certificates' function. This includes changes to use a 'base_path', 'description', 'configuration', and 'expression_file_paths' instead of 'country', 'environment', and 'certificates'. Additionally, the regex pattern used in the workflow to check the pull request title has been corrected to improve the matching functionality.

* Remove 'TRUSTSTORE' and 'TLS' from config

The configurations 'TRUSTSTORE' and 'TLS' were removed from the Brasil/dev/config.env file. These changes are considered necessary for the new application's environment setup and reducing unnecessary complexity.

* Update reference in GitHub workflow

The reference to an action file in the GitHub workflow named 'update_bucket.yml' was updated. Specifically, the line was changed from using the 'refatorando' branch to now use the 'main' branch in the repository. This change ensures that the workflow uses the latest stable version of the action.

* Add directory.properties file and update GitHub workflow

A new directory.properties file was added with the keystore and truststore values. This change allows the GitHub workflow, specifically within the update_bucket.yml config file, to reference this properties file. The benefits of this include improved file organization and easier maintainability.

* Update title regex in update_bucket workflow

The title regular expression in the GitHub Action workflow 'update_bucket.yml' has been updated. The previous version was constraining the titles with the term 'certificates' which has now been removed to allow for more generic titles.

* Update secrets inheritance in workflow

The position of 'secrets: inherit' line in the 'update_bucket.yml' GitHub workflow configuration has been adjusted. Now, secrets inheritance is declared earlier in the 'sync_certificates' block. This rearrangement is aimed at improving the workflow's readability and maintainability.

* Update sync certificates to specific ref

The action for the sync_certificates job in the update_bucket workflow has been updated to use a specific ref (6-refatorar-action) instead of the main branch. This change will ensure our pipeline always uses the specific version we have tested and verified, improving reliability.

* Update action reference in workflow

The commit updates the reference of the action used in the 'sync_certificates' job in Github Workflow. The reference has changed from '6-refatorar-action' to 'main', ensuring that the current stable version of the action is used for syncing certificates.

* Add new certificates to keystore and truststore

The commit adds two new certificates to the keystore and the truststore. Specifically, it includes a new certificate for 'escola.org' in the keystore and one for 'posto' in the truststore. These certificates are expected to serve as authoritative entities in their respective domains.