country Brasil environment dev delete (#13)
* Update GitHub Actions workflow for branch push

This commit modifies the GitHub Actions workflow (sync_bucket.yml) to include actions on 'push' events for the 'main' branch. It also restricts 'verifier' and 'planning' jobs to only execute on 'pull_request' events. Last, it removes a dependency on the 'planning' job for the 'push_to_main' job.

* Update action references in workflow file

The workflow file, sync_bucket.yml, has been updated to reference actions from the main branch instead of the 'feature/planning' branch. This change is applicable for both 'planning' and 'push_to_main' stages within the workflow.