Update GitHub workflows to use environment variables
The verifier job has been removed and incorporated into the planning job instead. Environment variables are now utilized for assigning the country and environment, which replaces the verifier's output previously used for these specific details. Secrets have been added to the 'planning' and 'apply' steps to manage AWS access and the 'apply' step now depends on the 'planning' job to ensure correct order of execution. The environment variables are also saved in an artifact file for use in the 'apply' job.