(Brasil dev) Add VIRTUAL_PATH to Brasil dev config file (#18)
* Add VIRTUAL_PATH to Brasil dev config file

The Brasil dev configuration file has been updated with the addition of a new environment variable, "VIRTUAL_PATH". This will specify the precise location in the bucket where files will be stored or accessed.

* Update pull request title validation regex

The pull request title validation regex in the GitHub workflow has been modified. This change allows underscores in addition to alphanumeric characters and spaces in the title.

* Update VIRTUAL_PATH in config.env

The VIRTUAL_PATH in the config environment file (config.env) for the Brazilian development environment has been updated. The previous value "123/456/789" was replaced with the new path "brasil/dev/certificates".